add s to project1111111111

+++ app/Http/Controllers/RoleController.php
@@ -3,18 +3,43 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\RoleRequest;
+use App\Http\Requests\UpdateRoleRequest;
 use App\Models\Permission;
 use App\Models\Role;
+use App\Rules\UniqueOtherSelf;
 use Illuminate\Http\Request;
 
 class RoleController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
         $this->authorize('isAccess',Permission::query()->where('title','list_roles')->first());
 
-        return view('roles.index',[
-            'roles'=>Role::all(),
+        $field = $request->get('field');
+        $phrase = $request->get('phrase');
+
+        if ($request->has('field') && $field == 'all') {
+            $roles = Role::query()->where('title', 'like', "%{$phrase}%")
+                ->where('id','!=', 1)
+                ->orWhere('description', 'like', "%{$phrase}%")
+                ->orderBy('id')->paginate();
+        } else if ($request->has('field') && !empty($field) && !empty($phrase)) {
+            $roles = Role::query()->where($field, 'like', "%{$phrase}%")
+                ->where('id','!=', 1)
+                ->orderBy('id')->paginate();
+        } else {
+            $roles = Role::query()->where('id','!=', 1)->paginate();
+        }
+
+        $fields = [
+            'title'=>'عنوان نقش',
+            'description'=>'توضیحات',
+            'all'=>'همه موارد',
+        ];
+
+        return view('pages.users.roles.index',[
+            'roles'=>$roles,
+            'fields'=>$fields
         ]);
     }
 
@@ -22,7 +47,7 @@ class RoleController extends Controller
     {
         $this->authorize('isAccess',Permission::query()->where('title','create_role')->first());
 
-        return view('roles.create',[
+        return view('pages.users.roles.create',[
             'permissions'=>Permission::all(),
         ]);
     }
@@ -31,7 +56,6 @@ class RoleController extends Controller
     {
         $this->authorize('isAccess',Permission::query()->where('title','create_role')->first());
 
-//        dd($request->all());
         $role = Role::query()->create([
             'title'=>$request->get('title'),
             'description'=>$request->get('description'),
@@ -44,20 +68,29 @@ class RoleController extends Controller
     {
         $this->authorize('isAccess',Permission::query()->where('title','edit_role')->first());
 
-        return view('roles.edit',[
-            'role'=>$role,
-            'permissions'=>Permission::all(),
-        ]);
+        if ($role->id != 1 && $role->title != 'admin') {
+            return view('pages.users.roles.edit', [
+                'role' => $role,
+                'permissions' => Permission::all(),
+            ]);
+        }
+        return redirect(route('roles.index'));
     }
 
-    public function update(RoleRequest $request,Role $role)
+    public function update(UpdateRoleRequest  $request,Role $role)
     {
         $this->authorize('isAccess',Permission::query()->where('title','edit_role')->first());
 
-        $role->update([
-            'title'=>$request->get('title'),
-            'description'=>$request->get('description'),
-        ]);
+        $this->validate($request,
+            ['title' =>new UniqueOtherSelf('roles','title',$role->id,"عنوان نقش تکراری است لطفا عنوان دیگری وارد کنید.")]
+        );
+
+        if ($role->id != 1 && $role->title != 'admin') {
+            $role->update([
+                'title' => $request->get('title'),
+                'description' => $request->get('description'),
+            ]);
+        }
 
         $role->permissions()->sync($request->get('permissions'));
         return redirect(route('roles.index'));
@@ -67,8 +100,13 @@ class RoleController extends Controller
     {
         $this->authorize('isAccess',Permission::query()->where('title','delete_role')->first());
 
-        $role->permissions()->detach();
-        $role->delete();
+        if ($role->id != 1 && $role->title != 'admin') {
+
+            $role->permissions()->detach();
+            $role->delete();
+
+        }
+
         return redirect(route('roles.index'));
     }
 }
